Add files via upload
This .txt file describes the initial steps I took to obtain, clean, and begin analyzing 
some public data on county-wide business patterns, restricting my focus to New York 
State. In further updates, I will commit an R script that does these things and more. 
In particular, I will also look at other years besides 2016, to see trends that emerge.
The point of this version is simply to develop the protocol for obtaining the data
and preparing it for analysis.